+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.0.7] - 2023-07-12
+
 ### Fixed
 - Fixed getting DAO list from core storage service
 - Fixed missed params in DAO and Strategy models